This is a simple login form made using html and css.
The login form is basically the same as the everyday login form found on most webpages. The css file in the project was generated by sass a css framework. It is a static login form and no links on it are active. This commit has only the mobile view of the login form, the tablet and desktop views will be committed, if God wills, in the coming days.